refactor(sales-app): change type of store field

+++ sales-app/app/src/main/java/com/peyess/salesapp/model/store/FSOpticalStore.kt
@@ -75,7 +75,7 @@ data class FSOpticalStore(
     @Keep
     @JvmField
     @PropertyName("days_to_take_from_store")
-    val daysToTakeFromStore: Float = 0f,
+    val daysToTakeFromStore: Int = 0,
 
     @Keep
     @JvmField
